Fix not casting starsurge Properly.
Fixed by looking back into the cast log.
changed the eclipse function to "player.eclipseRaw". Because why not :)
And removed trailing spaces.

Maybe you should add  { "#Empowered Augment Rune", "!player.buff(175457)" }, somewhere. :) (I add it to mine every time I update.)
Also is there anyway to fix the interrupts interrupting the players cast when the spell is on cooldown? I tried adding "!spell.cooldown" and that actually gets called but it still interrupts the player when it is on cooldown. We might have to replace InterruptsAt for that or use InterruptAt.