Quicker and more efficient.
Added FormatTime so it tells the time in hours, minutes and seconds.
Fix a Looted Item counter. Added a predicted average counter for it.
Added a cooldown for casting fishing because it seemed to double cast after applying the worm.
Changed the way some things work. (Mainly to stop interrupting the fishing.)
Moved tables out of functions to increase efficiency.